Fix a potential infinite loop, size_t will never be negative

+++ core_lib/managers/viewmanager.cpp
@@ -207,7 +207,7 @@ void ViewManager::scaleUp()
 
 void ViewManager::scaleDown()
 {
-    for (size_t i = mZoomLevels.size() - 1; i >= 0; --i)
+    for (int i = mZoomLevels.size() - 1; i >= 0; --i)
     {
         if (mZoomLevels[i] < scaling())
         {
